+++ NEWS.md
@@ -384,7 +384,7 @@ Excluding feature bug fixes.
 
 ### JIT
 
-* The default `--jit-max-cache` is changed from 100 to 10000.
+* The default `--mjit-max-cache` is changed from 100 to 10000.
 
 * JIT-ed code is no longer cancelled when a TracePoint for class events
   is enabled.
@@ -392,7 +392,7 @@ Excluding feature bug fixes.
 * The JIT compiler no longer skips compilation of methods longer than
   1000 instructions.
 
-* `--jit-verbose` and `--jit-warning` output "JIT cancel" when JIT-ed
+* `--mjit-verbose` and `--mjit-warning` output "JIT cancel" when JIT-ed
   code is disabled because TracePoint or GC.compact is used.
 
 * `RubyVM::MJIT` is renamed to `RubyVM::JIT`. [[Feature #17490]]